Fix dropping foreign key multiple times with test
In some cases the Comparator class returns multiple drops for the same foreign key.
Specifically, in case you have two tables, A & B, with A having a foreign key FK
referencing B, if you drop table B, the resulting diff show this FK twice,
once on the diff->orphanedForeignKeys as we're deleting B, and another on
diff->changedTables as table A is being modified. As a result of this you get the
DROP FOREIGN KEY instruction twice in the final SQL.