added the possibility to use column names as keys for the types in Connection insert() and update() methods
The Connection::insert() and Connection::update() methods take an array
of types as their third argument:

$conn->insert('foo', array('foo' => 'foo', 'date' => new \DateTime()),
array(2 => 'datetime'));

But without looking at the implementation, I'm sure you would have
written the following instead:

$conn->insert('foo', array('foo' => 'foo', 'date' => new \DateTime()),
array('date' => 'datetime'));

But unfortunately, that's does not work. This commit add support for
this, and of course, the old way still works.