Add method ResultStatement::rowCount()
Sometimes it is required to get the number of rows in a result before
fetching the data itself. PDO provide PDO_Statement::rowCount() for
this, wich is now be added to the ResultStatement interface.
This will only affect the class ArrayStatement, because it is the only
statement implementation, that does not provide rowCount().