ORACLE, (INSTANCE_NAME = XXXXX)
Hello, first sorry for my English. 
I need to put in the ORACLE Connection string to the parameter: (INSTANCE_NAME = XXXXX) 

I was reading the source code at github and the latest version does not include this possibility in the method getEasyConnectString. 

Could add to the next version an element in the array of parameters, eg $params['instance_name'] and concatenating that value to the Connection string? 

Thank you, greetings 
Facundo Panizza