Fixes issue (DBAL-1057) where simply initiating the platform may cause an error
For example, if you have no database, then previously, we would try to connect,
causing an error. But in this case, we can simple fall back to the default platform.
This is exactly what the previous version did (see sha: 3176f51d1426022d305c6531a9b9bc93a868bddd),
and thus this adds the nice fallback we had before, without removing the flexibility
of setting your exact server version or determining it once there *is* a connection