Changing implementation to *try* to connect
For me, this makes more sense. Without this, if getDatabasePlatform() is called
consistently before connection(), then the platform will *never* connect to
get the version and use the *best* platform - it will always use the fallback.
This *does* use the connection, but doesn't explode for the edge cases when
the connection information is invalid.