Strip leading slash of databasename from URL
When using a custom driver via `driverClass` in all (some?) cases one cannot set
the scheme properly. In the previous implementation when the scheme is missing the
`DriverManager` leaves the leading slash in the path, because it silently assumes,
that this is a SQLite-connection. With custom drivers this leads to invalid database
names.