DBAL should be able to see more type of index in DB
For example:
We have a query to create index like this:

CREATE INDEX ON films ((lower(title)));

This index would not be seen by DBAL because the old approach will
try to find a related column with this index but since this index
use expression '(lower(title))', there is no real column associated
with it.

The new approach will get all expressions from create index query
using pg_get_indexdef regardless of they are expressions or real
column.

In this way DBAL won't complain where an expression-based is used
when running scheme update command.