Adds default namespace to schema.
Some databases supports namespaces and has default one. For example
Postgres has default "public".

When schema created from classes, which has table names without
namespaces, but this namespace is assumed, we should add it explicitly,
if it not exists yet.

It will help in diffs between real DB and config shemas.

This is 1st part of fix DBAL-1168