unit tests for oracle type consistency.
add unit tests that try to generate a table with a column type for every
available doctrine data type. inject the data type information from the
platform's CreateTableSQL back through the schema manager to confirm that
the types are consistent. Any types that are not implemented by the platform
are marked as skipped.

This only tests Oracle types for now, but I've implemented an
AbstractSchemaManagerTest as I think that the principle can be applied to
other platforms as well.