Added an interface for drivers for platforms that support deactivation of foreign key checks
MySQL supports the deactivation of foreign key checks. This is helpful if one wants to purge the database. Usually Doctrine tries to figure out in which order tables have to be cleared. This must not be done if the platform supports the deactivation of foreign key checks.
This feature may be used in the ORMPurger of doctrine/data-fixtures.