Array Type - Empty String (NOT NULL) support
Setting a column as Array and NOT NULL for a table with existing data - database (at least MySQL) will set that column as '' (empty string) for existing rows.
And fetching those rows, Doctrine throws a ``Converstion Exception`` with message ``Could not convert database value "" to Doctrine Type array``

This PR fixes that by assuming empty string as empty array.