Automatic escaping of default values
It is now possible to use `\` in non-escaped default values for a column,
and it will be automatically escaped for platforms that need it.

Previously this lead to a confusion when diffing actual and expected
schema leading to perpetual out of sync schema.

Before:

```php
/**
 * @ORM\Column(options={"default" = "Foo\\Bar"}))
 */
private $name;

```

After:

```php
/**
 * @ORM\Column(options={"default" = "Foo\Bar"}))
 */
private $name;

```

And the result in database will be a default value exactly as written in
the annotation, that is `Foo\Bar`.