allow creating PRIMARY KEY AUTOINCREMENT fields for sqlite
If a PRIMARY KEY field is defined in sqlite without autoincrement then
keys might be reused when rows are deleted, explicitly settings autoincrement
will prevent this.

Autoincrement is not enabled by default due to additional overhead introduced
with autoincrement bookkeeping and should only be enabled when the additional
uniqueness is required.