Allow getCustomSchemaOption 'using' to be used
Allow getCustomSchemaOption 'using' to be used as an additional sql to an alter column type change.  This is to allow a way of adding custom SQL that can be used to convert from one type to another.  Fixes: https://github.com/laravel/framework/issues/28963 with `syntax ->change()->using('special_price::int::numeric(12,4)')`.