When renaming a primary key, and at the same time changing some columns the Comparator does not detect the name correctly.
For databases as MySQL where primary keys do not have a name, this does not matter.
For database as PostgreSQL where also primary keys have a name, the generated SQL will try to drop a column that does not exists (It will try to drop the index with the new name).