Add getAggregateConcatExpression to platform
This returns the expression for aggregating values by concatenating them
with a separator between the individual values. This corresponds to
something like GROUP_CONCAT/LIST/LISTAGG/STRING_AGG. Since there does
not really seem to be a standard, the function is generically named and
does not have a default implementation in AbstractPlatform.