SQLite: Fix wrongly detected reference constraint name on schema change
If there is no constraint name defined the listTableForeignKeys should not invent a new one as a numerical id has no deeper meaning and makes schema changes adding information that was not there before.
Using `null` is an apropriate value as its already in use when handling not existing fk names.