Add support for read-only SQLite databases
Since PHP 7.3 it's posible to open SQLite databases in read-only
mode[1]. I figured that it makes sense to add a driverOption for
`pdo_sqlite` to support this.

The idea behind it is to declare `read_only: true` in e.g. a sf-bundle
config without having to use the constants in the config files.

[1] https://github.com/php/php-src/commit/ed2f6510da7c68b5690c60344cbb9fe73241592a