CLOB type for ORACLEDB included.
For oci there is a difference in the types BLOB and CLOB, the class OCI8Statement treated only the BLOB fields. For this, a new type was created in the ParameterType class, which in the OCI8Statement class is converted to OCI_B_CLOB.
For parameter bind with type CLOB, ParameterType :: CLOB must be used.
Fixes #3290