Exclude duplicate SQL statements from migrations (schemadiff->toSql)
This is a bug fix for the command doctrine:migrations:diff. A simple way to reproduce the buggy behaviour is to 1) create two entities and a unidirectional one-to-one relationship between them, 2) update the database schema, 3) rename the table that is foreign in the key constraint, 4) run migrations:diff. The result is a migration that always fails because of a duplicate drop statement for the same foreign key.

Things taken into consideration in the bug fix:
 - The to-be-orphan foreign key must be dropped before the table is dropped.
 - A new key cannot be created before the new table has been created.
 - When the new key is created, the old key must not be dropped again because it has already been dropped.

Regarding the correctness of the result, it is the latter SQL drop statement that should be removed to achieve a executable migration and desired schema update. It is currently added in the code that adds ALTER TABLE statements based on a TableDiff but unaware of the other parts of SchemaDiff. So I reckon the best place for the fix is in SchemaDiff.php where all the SQL statements are gathered. I've added new condition for the incoming ALTER TABLE statements: only include them if not yet included. This condition would also apply to other kind of statements but I haven't come across any bugs in any other SQL statements.