Add error handler around oci_fetch_array to screen for oracle warnings
The error handler screens for oracle warnings from php oci8 driver
and converts them to a truncated query result exception. This also adds a test
that demonstrates the issue where a result set larger than the oci8
default prefetch value is queried, but a truncated
row set is returned due to the dataset being invalidated during the
fetch sequenece.