BUGFIX: sqlite has no knowledge about longblob
This fixes a bug where webapps, who are trying to create a longblob field in sqlite, are going wrong. The bugfix is based in the comment by @Naziak on a similar issue with MySQL in the Nextcloud server repository (https://github.com/nextcloud/server/issues/25748#issuecomment-784736577).
The bugfix has successfully been tested on my personal nextcloud instance.
I decided to not create a issue on github as why bother people if typing this message takes 10x more time than fixing the actual bug?