Introduce platform family abstraction
The possible return values of AbstractPlatform::getName() need to stay
very stable because consumers are relying on those in a lot of
scenarios, probably even more than they rely on instanceof, since
doing so for e.g. PostgreSQLPlatform no longer works with DBAL 3.
Introducing the possible values as constants of a new type will:
- make these strings explicitly part of our public API;
- help consumers understand better where these strings are defined, and
  hopefully make them add a dependency to doctrine/dbal when they need
  to;
- avoid bugs thanks to IDE auto-completion (less risk of typing
  posgresql instead of postgresql);
- avoid more bugs thanks to static analysis provided by the phpdoc.