Streamlines identity sequence name with PostgreSQL
Using "autoincrement" option creates a sequence with the table name
(e.g. test) and _SEQ as suffix, i.e. test_SEQ. If unquoted, Oracle turns
this to TEST_SEQ but if the table name is quoted, it will be "test_SEQ".

This change adds lower case "_seq" as suffix so test_seq will be still
turned to TEST_SEQ by Oracle but quoted "test_seq" stays as is. Quoted
table names will now create the same sequence names as PostgreSQL does so
no special handling of "test_SEQ" in Oracle and "test_seq" in PostgreSQL
is necessary any more.