Fix SQLite date arithmetic for dynamic intervals
Fixes #4426, #4462.

I am really only personally concerned about handling seconds/minutes/hours
correctly, but figured it made sense to try and fix weeks and quarters as well
since I'm in the area. I'm open to reverting those changes or instead simply
asserting if `$unit` is WEEK or QUARTER and the $interval is not numeric, so at
least clear error messages are produced.

It looks like #3019 added a test case to `DataAccessTest` to check that date
arithmetic with a column reference worked. That change also enabled the interval
to be parameterized, so I also added test cases to ensure that behavior keeps
working, for all intervals.