Extended Connection to use different Type Registry between instances
Currently, every connection is using the same Type registry with the same static types, this can be problematic when we need to override a built-in type for one connection but not for the other.

A good example is a mysql and a mssql connection where the datetime types needs to be overridden for the mssql connection.

Another example could be a multi-tenant application where different clients use custom types with different configurations.