Adding the constant PARAM_ASCII_ARRAY to Doctrine\DBAL\Connection
This fix enable to use ASCII_STRING as type for Arrays of strings.
This new parameter was added in version 2.11.x but unfortunally we don't have a way to use in a array.

Also added it in the check for Param array in SQLParserUtils.php so we can handle it there too