Improve performance of Oracle constraint query
This patch slightly qualifies the query used to obtain Oracle constraint
metadata so that it can make better use of indexes on the constraint
view's underlying tables. There is no change in functionality or output,
just an improvement in speed. On my Oracle system, this patch provides a
4X improvement in speed, dropping a scan of about 400 tables from a
painfully slow 12 minutes to a slightly less painfully slow 3 minutes.