Failing test case: comparator yields invalid diffs when `Column#$_columnDefinition` is set
When a `Column` is compared against another `Column` with a `$_columnDefinition`, it yields invalid
results, because it generates a diff, but does not declare that the `columnDefinition` is the changed
property.

Before `3.3.1`, this behavior worked fine, but https://github.com/doctrine/dbal/pull/5220 introduced
a regression that led to the schema comparison changing behavior here.