When using the comparator via magic methods, reset the comparator state between calls
Before this change, the comparator would have internal dirty state before `Comparator#doCompareSchemas()` was called: with
this change, the state is reset before each call, giving us better long-term stability.

This change is still unclear: no tests currently fail with it on SQLite, but it is possible that the `Comparator` design
was accounting for more complexity on more exotic platforms.

This may not necessarily be the correct fix to the https://github.com/doctrine/dbal/pull/5220 regression.