MySQL: add required length for primary key with column of TextType
This fixes the mysql error "SQLSTATE[42000]: Syntax error or access violation: 1170 BLOB/TEXT column 'document_id' used in key specification without a key length" when the length is missing which is only required for index column of TextType.

Length for column with TexType will be set to 255 if no length is set or the values of the new parameter $options["lengths"] are used to set this value.