Deprecate ColumnDiff::$oldColumnName and ::getOldColumnName()
The $oldColumnName property and the corresponding constructor parameter
were relevant until https://github.com/doctrine/dbal/pull/220 (2.4.0)
which introduced the $fromColumn property and the corresponding
constructor parameter. Now they are redundant.

The same applies to getOldColumnName(). The $fromColumn property
contains all the properties of the old column including the name.