MariaDb1043. Detect the need to migrate JSON columns to native JSON.
MariaDb JSON columns are marked as json with a DC2Type comment hint.

The schema manager maps such columns to a json doctrine type which
matches the metadata whether the column was declared as LONGTEXT (by the
previous MariaDb1027 platform) or JSON. This means the comparator will
not detect that a column should be upgraded from LONGTEXT to native JSON
and no migration would be generated.

Therefore, during column introspection check that the data type of the
introspected columns match that which should be set for any doctrine
types inferred from DC2Type comments. Set a flag (a platformOption
for the relevant column) where it does not.

Check whether the flag is set when the comparator diffs the expected
and actual tables and mark a column type difference where the flag is
set so a migration is generated.