DB2Platform. Add variable precision datetime support.
Add precision to TIMESTAMP etc columns where column specification
includes a scale. Where no scale is specified, default remains
TIMESTAMP(0).

DB2 supports precisions of between 0 and 12 for datetime columns but not
time columns which are to the nearest second only. Since php datetime
objects support a maximum precision of only 6, this is the maximum useful
precision although datetime columns with a scale of 12 can still be set.

The DB2 schema manager cannot infer a doctrine datetimetz type from a
database column as the type lacks a DC2Type comment. They are introspected
as datetime columns. The comparator then identifies a difference
between doctrine and database schemas for datetimetz types. This error
is ignored in the VariableDateTimePrecisionIntrospectionTest as is not
due to adding variable precision datetime support.

ISSUE WITH SUPPORTING VARIABLE PRECISION TIME

DB2 uses TIME columns for TimeTypes and TIME does not support fractional
seconds. Switching to TIMESTAMP would require changing the format string
to specify a fixed date.

This means variable precision time columns cannot be implemented in a
backwards compatible way (i.e. without converting all pre-existing TIME
columns in the database) because TimeType::convertToDatabaseValue() has
no knowledge of the column, the details of which would be needed to choose
the correct format string for conversion.