OraclePlatform. Add opt-in variable precision datetime support.
Add precision to TIMESTAMP columns where column specification includes a
scale. Where no scale is specified, default remains TIMESTAMP(0) or
TIMESTAMP(0) WITH TIME ZONE.

Oracle supports from 0-9 fractional seconds for TIMESTAMP columns only.
On the Oracle platform, variable precision is therefore only available
for DateTimeType and DateTimeTzType and not TimeType.

Oracle sets the string format of timestamps in session variables
(NLS_TIMESTAMP_FORMAT and NLS_TIMESTAMP_TZ_FORMAT). Changing these
settings to support high precision timestamps means Oracle transmits a
changed string representation of datetime objects to the application
which is a BC break.

Using high precision datetimes is therefore an opt-in feature.

Add a new platform, OracleHptPlatform, which supports high precision
timestamps.

Add a boolean driverOption ('high_precision_timestamps') detected by
the Oracle drivers which, if the option is true, set a flag used by the
AbstractOracleDriver to determine which platform instance to return (the
platform supporting high precision timestamps or the legacy platform).

The driver option is also detected by the InitializeSession middleware
and OracleSessionInit post-connect event handler, in both cases to
determine the appropriate NLS_TIMESTAMP_ format strings.

This ensures the NLS_TIMESTAMP_ format strings and php datetime format
strings set by the platform match.

Use '.FF6' as the high precision specifier as it matches php's '.u' format
string. The maximum practical time resolution on the Oracle platform is
therefore microsecond. Trailing zeroes will be added in the database
where a higher precision is specified by the column but these will not
be available to doctrine.

If the high_precision_timetamps driverOption is set to false or unset,
precision can still be added to datetime columns but the highest
resolution timestamp that can be represented will remain one second.

The schema manager cannot identify a doctrine TimeType from a database
column because DATE is used and this maps to DateType without a DC2TYpe
comment. This introspection error is ignored in the relevant test as is
not due to adding variable precision datetime support.

ISSUE WITH SUPPORTING VARIABLE PRECISION ON TIME COLUMNS

Oracle uses DATE columns for TimeTypes and DATE does not support
fractional seconds. Switching to TIMESTAMP would require changing the
format string to specify a fixed date for TimeTypes.

This means variable precision time columns cannot be implemented in a
backwards compatible way (i.e. without converting all pre-existing DATE
columns in the database) because TimeType::convertToDatabaseValue() has
no knowledge of the column, the details of which would be needed to
choose the correct format string for conversion.

** Also fixes php format strings for datetimes including timezones **

The datetime format Oracle uses for conversion of strings to times is
specified in NLS_TIMESTAMP_TZ_FORMAT.

It has historically been set to 'YYYY-MM-DD HH24:MI:SS TZH:TZM', i.e.
with a space between time and timezone.

The php format string was historically set as 'Y-m-d H:i:sP', i.e.
without a space between time and timezone.

This means that times with negative timezone differences are read from
the database incorrectly as times with positive timezone differences.

This patch conforms the php format string to the NLS_TIMESTAMP_TZ_FORMAT.